requirements.txt
amqp==2.1.4
appdirs==1.4.0
asgi-redis==1.0.0
asgiref==1.0.0
attrs==16.3.0
autobahn==0.17.1
billiard==3.5.0.2
cached-property==1.3.0
celery==4.0.2
channels==1.0.3
colorama==0.3.7
constantly==15.1.0
coreapi==2.2.4
coreschema==0.0.4
coverage==4.3.4
daphne==1.0.3
decorator==4.0.11
defusedxml==0.5.0
Django==1.11
django-cors-headers==2.0.2
django-extensions==1.7.7
django-extra-fields==0.9
django-filter==1.0.4
django-markdown-deux==1.0.5
django-rest-swagger==2.1.1
djangorestframework-csv==2.0.0
djangorestframework==3.6.2
docker==2.1.0
docker-compose==1.11.1
docker-pycreds==0.2.1
dockerpty==0.4.1
docopt==0.6.2
fcm-django==0.2.7
googlemaps==2.4.6
incremental==16.10.1
itypes==1.1.0
Jinja2==2.9.5
jsonschema==2.6.0
kombu==4.0.2
markdown2==2.3.3
MarkupSafe==0.23
model-mommy==1.3.2
msgpack-python==0.4.8
oauthlib==2.0.1
olefile==0.44
openapi-codec==1.3.1
packaging==16.8
paypalrestsdk==1.12.0
pep8==1.7.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.0.0
prompt-toolkit==1.0.13
psycopg2==2.6.2
ptyprocess==0.5.1
pycodestyle==2.3.1
pyfcm==1.2.3
Pygments==2.2.0
PyJWT==1.4.2
pyparsing==2.1.10
python-dateutil==2.6.0
python-google-places==1.4.0
python3-openid==3.1.0
pytz==2017.2
PyYAML==3.12
redis==2.10.5
requests==2.13.0
requests-oauthlib==0.8.0
rest-condition==1.0.3
service_identity==16.0.0
simplegeneric==0.8.1
simplejson==3.10.0
six==1.10.0
stripe==1.51.0
social-auth-app-django==1.1.0
social-auth-core==1.2.0
texttable==0.8.7
traitlets==4.3.2
Twisted==17.1.0
txaio==2.6.1
uritemplate==3.0.0
vine==1.1.3
wcwidth==0.1.7
websocket-client==0.40.0
zope.interface==4.3.3
geopy==1.11.0
nltk==3.2.1
